+++ CHANGELOG.md
@@ -1,2 +1,3 @@
 # 0.1.0 (Unreleased)
 * Initial release
+* Add `volumes` for host-to-container bind mounts.